Update 2021_09_20_124499_create_habitaciones_table.php
Se modifico el campo de la tabla para el numero de habitacion de tinyInteger a Integer

+++ backend/api/database/migrations/2021_09_20_124499_create_habitaciones_table.php
@@ -14,7 +14,7 @@ class CreateHabitacionesTable extends Migration
             $table->foreignId('hotel_id');
             $table->foreignId('tipo_id');
             $table->tinyInteger('piso');
-            $table->tinyInteger('numero');
+            $table->Integer('numero');
             $table->bigInteger('precio');
             $table->boolean('minibar');
             $table->tinyInteger('capacidad');
